+++ packages/intents-sdk/src/constants/poa-tokens-migrated-to-omni-bridge.ts
@@ -1,14 +1,9 @@
 import { ChainKind } from "@omni-bridge/core";
 
 /**
- * These tokens can be routed through either PoA Bridge or Omni Bridge
- * depending on the SDK configuration. Use the `routeMigratedPoaTokensThroughOmniBridge`
- * feature flag to enable Omni Bridge routing for these tokens.
+ * These tokens originaly have been PoA Bridge tokens but have been migrated to Omni bridge.
  */
-export const POA_TOKENS_ROUTABLE_THROUGH_OMNI_BRIDGE: Record<
-	string,
-	ChainKind
-> = {
+export const POA_TOKENS_MIGRATED_TO_OMNI_BRIDGE: Record<string, ChainKind> = {
 	"sol-57d087fd8c460f612f8701f5499ad8b2eec5ab68.omft.near": ChainKind.Sol, // BOOK OF MEME,
 	"sol-c58e6539c2f2e097c251f8edf11f9c03e581f8d4.omft.near": ChainKind.Sol, // OFFICIAL TRUMP
 	"sol-b9c68f94ec8fd160137af8cdfe5e61cd68e2afba.omft.near": ChainKind.Sol, // WIF
@@ -21,4 +16,8 @@ export const POA_TOKENS_ROUTABLE_THROUGH_OMNI_BRIDGE: Record<
 	"sol-c634d063ceff771aff0c972ec396fd915a6bbd0e.omft.near": ChainKind.Sol, // SPX
 	"sol-d600e625449a4d9380eaf5e3265e54c90d34e260.omft.near": ChainKind.Sol, // MELANIA
 	"sol.omft.near": ChainKind.Sol, // SOL
+	"aptos.omft.near": ChainKind.Aptos, // APTOS
+	"aptos-34ee497f210c5a511e8d5b53bc56d75b63612bb5.omft.near": ChainKind.Aptos, // APTOS USDC
+	"aptos-88cb7619440a914fe6400149a12b443c3ac21d59.omft.near": ChainKind.Aptos, // APTOS USDT
+	"starknet.omft.near": ChainKind.Strk, // STARKNET
 } as const;